feat: Fixed lint issues, moved tests to test and added tools folder with dependencies
