+++ Makefile
@@ -7,8 +7,10 @@ INTEGRATION := jmx
 BINARY_NAME  = nri-$(INTEGRATION)
 GO_PKGS     := $(shell go list ./... | grep -v "/vendor/")
 GO_FILES    := ./src/
-GOTOOLS      =  github.com/axw/gocov/gocov \
-		github.com/AlekSi/gocov-xml \
+GOFLAGS			= -mod=readonly
+GOLANGCI_LINT	= github.com/golangci/golangci-lint/cmd/golangci-lint
+GOCOV           = github.com/axw/gocov/gocov
+GOCOV_XML		= github.com/AlekSi/gocov-xml
 
 all: build
 
@@ -18,42 +20,22 @@ clean:
 	@echo "=== $(INTEGRATION) === [ clean ]: Removing binaries and coverage file..."
 	@rm -rfv bin coverage.xml $(TARGET)
 
-tools: check-version
-	@echo "=== $(INTEGRATION) === [ tools ]: Installing tools required by the project..."
-	@go get $(GOTOOLS)
-	@curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s v1.33.0
-
-tools-update: check-version
-	@echo "=== $(INTEGRATION) === [ tools-update ]: Updating tools required by the project..."
-	@go get $(GOTOOLS)
-	@curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s v1.33.0
-
-deps: tools
-
-validate: deps
+validate:
 	@echo "=== $(INTEGRATION) === [ validate ]: Validating source code running golangci-lint..."
-	@./bin/golangci-lint run
+	@go run $(GOFLAGS) $(GOLANGCI_LINT) run --verbose
 
-validate-all: deps
-	@echo "=== $(INTEGRATION) === [ validate ]: Validating source code running golangci-lint..."
-	@./bin/golangci-lint run
-
-compile: deps
+compile:
 	@echo "=== $(INTEGRATION) === [ compile ]: Building $(BINARY_NAME)..."
 	@go build -o bin/$(BINARY_NAME) ./src
 
-compile-only: deps-only
-	@echo "=== $(INTEGRATION) === [ compile ]: Building $(BINARY_NAME)..."
-	@go build -o bin/$(BINARY_NAME) ./src
-
-test: deps
+test:
 	@echo "=== $(INTEGRATION) === [ test ]: Running unit tests..."
-	@gocov test -race $(GO_PKGS) | gocov-xml > coverage.xml
+	@go run $(GOFLAGS) $(GOCOV) test -race ./... | go run $(GOFLAGS) $(GOCOV_XML) > coverage.xml
 
 integration-test:
 	@echo "=== $(INTEGRATION) === [ test ]: running integration tests..."
-	@docker-compose -f tests/integration/docker-compose.yml up -d --build
-	@go test -v -tags=integration ./tests/integration/. -count=1 ; (ret=$$?; docker-compose -f tests/integration/docker-compose.yml down && exit $$ret)
+	@docker-compose -f test/integration/docker-compose.yml up -d --build
+	@go test -v -tags=integration ./test/integration/. -count=1 ; (ret=$$?; docker-compose -f test/integration/docker-compose.yml down && exit $$ret)
 
 # Include thematic Makefiles
 include $(CURDIR)/build/ci.mk
@@ -62,13 +44,13 @@ include $(CURDIR)/build/release.mk
 check-version:
 ifdef GOOS
 ifneq "$(GOOS)" "$(NATIVEOS)"
-	$(error GOOS is not $(NATIVEOS). Cross-compiling is only allowed for 'clean', 'deps-only' and 'compile-only' targets)
+	$(error GOOS is not $(NATIVEOS). Cross-compiling is only allowed for 'clean' target)
 endif
 endif
 ifdef GOARCH
 ifneq "$(GOARCH)" "$(NATIVEARCH)"
-	$(error GOARCH variable is not $(NATIVEARCH). Cross-compiling is only allowed for 'clean', 'deps-only' and 'compile-only' targets)
+	$(error GOARCH variable is not $(NATIVEARCH). Cross-compiling is only allowed for 'clean' target)
 endif
 endif
 
-.PHONY: all build clean tools tools-update deps validate compile test integration-test check-version
+.PHONY: all build clean validate compile test integration-test check-version
