+++ src/config_parse_test.go
@@ -15,9 +15,9 @@ func TestParseYaml(t *testing.T) {
 		file         string
 		expectedFail bool
 	}{
-		{"../tests/data/test-sample.yml", false},
-		{"../tests/data/test-sample-bad.yml", true},
-		{"../tests/data/test-sample-nonexistant.yml", true},
+		{"../test/data/test-sample.yml", false},
+		{"../test/data/test-sample-bad.yml", true},
+		{"../test/data/test-sample-nonexistant.yml", true},
 	}
 
 	for _, tc := range testCases {
@@ -258,7 +258,11 @@ func TestParseCollectionDefinition(t *testing.T) {
 		},
 	}
 
-	c, err := parseYaml("../tests/data/test-sample.yml")
+	c, err := parseYaml("../test/data/test-sample.yml")
+	if err != nil {
+		t.Errorf("Unexpected error: %s", err.Error())
+		t.FailNow()
+	}
 	domains, err := parseCollectionDefinition(c)
 	if err != nil {
 		t.Error(err)
@@ -273,7 +277,11 @@ func TestParseCollectionDefinition(t *testing.T) {
 
 func TestParseCollectionDefinition_Fail(t *testing.T) {
 
-	c, err := parseYaml("../tests/data/test-sample-bad2.yml")
+	c, err := parseYaml("../test/data/test-sample-bad2.yml")
+	if err != nil {
+		t.Errorf("Unexpected error: %s", err.Error())
+		t.FailNow()
+	}
 	_, err = parseCollectionDefinition(c)
 	if err == nil {
 		t.Error("Expected error")
