+++ test/integration/json-schema-files/jmx-schema.json
