+++ test/integration/jsonschema/jsonschema.go
