+++ tools/tools.go
@@ -1,0 +1,9 @@
+// +build tools
+
+package tools
+
+import (
+	_ "github.com/AlekSi/gocov-xml"
+	_ "github.com/axw/gocov/gocov"
+	_ "github.com/golangci/golangci-lint/cmd/golangci-lint"
+)
